Bump Magick.NET to fix vulnerability

+++ Demo-ImageMagick/QrCodeImageExtensions.cs
@@ -8,6 +8,7 @@
 //
 
 using ImageMagick;
+using ImageMagick.Drawing;
 
 namespace Net.Codecrete.QrCodeGenerator;
 
@@ -44,7 +45,7 @@ public static class QrCodeImageExtensions
         }
 
         var size = qrCode.Size;
-        var dim = (size + border * 2) * scale;
+        var dim = (uint)((size + border * 2) * scale);
 
         if (dim > short.MaxValue)
         {
